Staging for new branch of project. Please Work!
The goal of this project is to add a UI to the existing RPS game that
was created earlier. I am essentially praying that this whole git branch
stuff is going to work and that this will not mess things up but here
we go! I added a CSS file to add some more cusomtization to the UI
elements. Also removed some sections of code from the old JS file to
facilitate the new system. Now comes the task of refactoring the JS to
use DOM methods to manipulate the UI instead of playing the game in the
console.