Initialized UI, Reworked Logic for DOM Methods
I made some changes to the structure of the HTML to make the UI more
readable when it changes. I also made changes to the logic of the game
so that the player's choice is registered before the function for the
game is called. This is handled by adding event listeners to each button
that call the function on click. Then depending on the button that was
pressed (read via event.target.id), the main function for the game will
initialize with the players choice passed in as an argument.

Need to work on keeping track of the score and declaring a winner once
a player reaches a score of 5. I would also like to add some styling to
spruce things up a bit.