Completed Script and adjusted HTML
Completed the script for the game which now runs entirely off of the
button presses. Had some difficulty updating the scores and announcing
a winner. I need to remember to initialize variables outside of the
functions if I am going to need the values stored after the functions
are called. Made some HTML adjustments to make some calls easier and
remove some unused classes/ids.